+++ src/types/api/responses/authorize.ts
